+++ scripts/run-shellcheck.sh
@@ -13,6 +13,7 @@ targets=(
   lib/common.sh
   lib/config.sh
   lib/access.sh
+  lib/frappe.sh
   lib/support.sh
   lib/backup.sh
   lib/ssl.sh